+++ NEWS.md
@@ -1,5 +1,21 @@
+# QFeatures 1.7
+
+## QFeatures 1.7.1
+
+- refactor: imputation now adds a new assay instead of replacing values.
+
+## QFeatures 1.7.0
+
+- New Bioc devel version.
+
 # QFeatures 1.5
 
+## QFeatures 1.5.3
+
+- feat: aggregation by adjacency matrix
+- New `adjacencyMatrix,SummarizedExperiment` and
+  `adjacencyMatrix,QFeatures` methods using
+  `ProtGenerics::adjacencyMatrix` (available in version >= 1.27.1).
 
 ## QFeatures 1.5.2